requirements.txt
# --- Whisper environment ---
torch
torchvision
torchaudio
faster-whisper
sounddevice
soundfile
numpy

# --- RAG environment ---
faiss-cpu
requests
sentence-transformers
